戦績情報の登録

+++ src/main/java/model/User.java
@@ -94,7 +94,7 @@ public class User {
  	}
 	
 	public double getRate() {
-		return (this.win / this.draw);
+		return ((double)this.win / (double)this.draw);
 	}
 	
 }
